requirements.txt
anthropic>=0.43.0
Pillow>=10.0.0
python-dotenv>=1.0.0
